Auto Merge (#40)
* sign in

sign in page

* sign in page

create sign in page

* deleted unnecessary things

And got a new bug. @Freeze2112, Figma code only works on specific device screen. It doesn't work on different Device Screen.

* fixed issue

closes #35

* Sign Up page

Developing Sign Up page

* updated and error fixed signup page

fixed
- error in overflowed pixels
- not working photo
- some spelling and positioning widgets error on signup page

* insert function for demonstrating home and signIn page

* nothing just testing

* Revert "nothing just testing"

This reverts commit ebb16d82c26bed754b8bc4f2a6954cb5d69b711e.

* fixed the issues of not able to scroll

removed the Stack Widget and changed the structure of both Sign In and Sign Up page in order to be scrollable if needed

* Changed Everything according to Project structure

---------

Co-authored-by: Freeze2112 <hlapyae321@gmail.com>
Co-authored-by: Freeze2112 <111502931+Freeze2112@users.noreply.github.com>